pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "klisk"
version = "0.1.0"
description = "A framework for building AI agents programmatically"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.1",
    "typer>=0.9",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "websockets>=12.0",
    "watchfiles>=0.21",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
litellm = [
    "openai-agents[litellm]",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[project.scripts]
klisk = "klisk.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/klisk"]

[tool.hatch.build.targets.wheel.force-include]
"src/klisk/templates" = "klisk/templates"
"src/klisk/studio_dist" = "klisk/studio_dist"
"src/klisk/chat_dist" = "klisk/chat_dist"

[tool.pytest.ini_options]
asyncio_mode = "auto"
